+++ src/packages/HLE/src/Memory/NativeMemory.T.cs
@@ -88,7 +88,7 @@ public sealed unsafe partial class NativeMemory<T> :
         ArgumentOutOfRangeException.ThrowIfNegative(length);
 
         nuint byteCount = checked((uint)sizeof(T) * (nuint)(uint)length);
-        T* memory = (T*)NativeMemory.AlignedAlloc(byteCount, (nuint)sizeof(T));
+        T* memory = (T*)NativeMemory.AlignedAlloc(byteCount, (uint)sizeof(T));
 
         if (zeroed)
         {
@@ -122,7 +122,7 @@ public sealed unsafe partial class NativeMemory<T> :
             return;
         }
 
-        Debug.Assert(MemoryHelpers.IsAligned((void*)memory, (nuint)sizeof(T)));
+        Debug.Assert(MemoryHelpers.IsAligned((void*)memory, (uint)sizeof(T)));
         NativeMemory.AlignedFree((void*)memory);
     }
